+++ client/lib/ft4/index.ts
@@ -16,8 +16,20 @@ import {
   createWeb3ProviderEvmKeyStore,
   createInMemoryEvmKeyStore,
 } from "./authentication";
-import { AuthDescriptor, FlagsType, Account, authDescriptor } from "./accounts";
+import {
+  AuthDescriptor,
+  FlagsType,
+  Account,
+  authDescriptor,
+  GtvAuthDescriptor,
+} from "./accounts";
 import { createAmount } from "./asset/amount";
+import {
+  registerAccount,
+  addRateLimitPoints,
+  registerAsset,
+  mint,
+} from "./admin/admin-op-functions";
 import { createEvmKeyHandler, EvmKeyStore, FtKeyStore } from "./authentication";
 import { createInMemoryFtKeyStore } from "./authentication/ft/key-stores/in-memory";
 import { Session } from "./types";
@@ -26,7 +38,6 @@ import {
   TransferHistoryResponse,
   TransferHistoryType,
 } from "./accounts/transfer-history/types";
-import { GtvAuthDescriptor } from "./accounts";
 import { DecimalFormat } from "./asset/types";
 
 export {
@@ -54,6 +65,10 @@ export {
   createLocalStorageLoginKeyStore,
   createEvmKeyHandler,
   createAmount,
+  registerAccount,
+  addRateLimitPoints,
+  registerAsset,
+  mint,
 };
 
 export const ft = Object.freeze({